Field label is "Custom Association Type" (#341).
When editing an association definition the field to set the custom association type is properly labeled "Custom Association Type" instead of "Association Type" which was ambigous as such a field exists in the same form already. Furthermore this is consistent with the labeling used in the Type Editor (when selecting a type and pressing Edit).

Note: this field label is generated from the association definition meta-model. Remember that in the detail panel a field of a simple child topic is labeled with the name of the custom association type (instead of with the name of the child type), if one is involved. The custom association type is now associated to an association definition via an association of type "Custom Association Type" (instead of "Aggregation"). The association type "Custom Association Type" itself is the custom association type of the association definitions that attaches the meta-type "Association Type" as a child to the association types "Composition Definition" and "Aggregationn Definition" in order to define the concept of a Custom Association Type in the first place. That means that the meta-model is defined partly by itself. Phew... that was fun ;-)

CHANGES

Core API

1 new `AssociationDefinitionModel` convenience constructor that takes a Custom Association Type URI (but no ID, URI, or view config):
{{{
AssociationDefinitionModel(String assocTypeUri, String customAssocTypeUri,
                           String parentTypeUri, String childTypeUri,
                           String parentCardinalityUri, String childCardinalityUri)
}}}

See #341.