Core: more flexible declarative migrations (#778).
CHANGE

The JSON format for a declarative migration is extended.
A declarative migration can consit of either of these:
1. A 4-property **JSON object** (the 4 optional properties are: `topic_types`, `assoc_types`, `topics`, `assocs`).
2. A **JSON array** of 4-property JSON objects.

The 2nd option is new.
DM processes the 4-property JSON objects contained in the array in order.

Note: DM 4.6 is backward compatible with older migration files.
Your existing migration files are not required to be changed.

See #778.