Redefine Contacts model (#779).
The topic types Phone Entry and Address Entry are replaced by respective association types. As a consequence the intermediate Phone Entry and Address Entry topics disappear.

No migration is yet provided to convert existing Contacts data.

One new problem arises with Custom Association Types in conjunction with detail panel rendering: the field formerly labeled "Phone" (which holds the actual phone number) is now accidentally labeled "Phone Entry". That is because of the current rendering rule, which is: if a **simple** (child) field is connected via an association definition which has a Custom Association Type set, then label the simple field with the name of the Custom Association Type instead of the name of the child type. This rule was made with the "Day of Birth"/"Day of Death" example scenario in mind which involves 2 association definitions with the same child type -- "Date" -- but differentiated via the 2 custom association types "Day of Birth"/"Day of Death". In this case the 2 fields are expected to be labeled "Day of Birth"/"Day of Death" instead of just "Date".

The new Phone Entry model reveals that this rendering rule does not hold in general.

Note: with **composite** child types in contrast this problem does not occur as a composite box in the detail panel, e.g. Address, is never labeled. Only the simple fields are.

=> We might be required to conceive another rendering rule.

This let the question arise if it is possible at all to derive the generic detail panel rendering solely through the data model or if we are required to enhance the data model with view configuration data.

See #779.