Webclient: revert assoc type label rule (#341).
The name of a Custom Association Type is no longer used as a field label in the detail panel.

Commit [0dc005de] is reverted. See discussion in ticket:779#comment:3

This brakes the "Birthday" example scenario but at least works well with the Contacts model that is delivered with the DM Standard Distribution.

See #341.