Webclient: reveal specific association (#460).
When clicking a topic in the detail panel which has more than one association to the selected topic only **one** association is revealed (instead of all), nameley the one that corresponds to the clicked list entry. This reduces the chance for selecting the wrong association on the canvas afterwards.

Note: If there are more than one association the related topic is listed in the detail panel several times, along with the names of the respective association types. Thus you can reveal the particular association you want.

Thanks to Malte for reporting!

CHANGES

Webclient API:

1 changed method:
{{{
dm4c.do_reveal_related_topic(topic_id, action, assoc_type_uri)
}}}
New is the 3rd parameter (optional). If not specified no association type filter is applied.

See #460.