Core API: Compositions support topic refs (#786).
When updating child topics a Topic Reference can also be used for childs which have an underlying Composition Definition. (Formerly topic references were only supported in conjunction with Aggregation Definition.)

This widens the usage context of the following APIs:

Core API:

* `ChildTopicModel`'s `putRef()` and `addRef()` methods
* `ChildTopic`'s `setRef()` and `addRef()` methods

REST API:

* The `ref_id:` and `ref_uri:` prefixes used in update requests

Note: when a topic reference appears in an operation for updating childs defined as "Composition One" the formerly assigned child topic (if there is one) is //deleted//. In contrast a child defined as "Aggregation One" is //not// deleted in that case but only the assignment (=association) is. This reflects the fact that a "compositioned" child can't exist without a parent relationship. An "aggregated" child in contrast can.

See #786.