Webclient: fix form processing (#341).
E.g. creating a Person and leaving the Phone field empty works again.

Note: when the Phone field is empty no empty Phone topic will be created (resp. if the field is emptied the Phone topic will be deleted), see #171. That means that there is no Phone Entry association either in this case. As a consequence no Phone Label ("home", "work", ...) can be stored. There can be no Phone Label without a Phone number.

This is a generic pattern. Form input bound to a custom association is ignored/deleted if the child topic is not created resp. deleted.

See #341.