Webclient: custom assoc type label rule (#341).
The detail panel renders the name of the custom association type as the child topic's field label (instead of the child type's name) if all conditions are met:

    1. the child type is simple
    2. a custom association type is set
    3. the custom association type is simple

New is the 3rd condition.

This rule is considered generally applicable and to work well with all kinds of application models.
At least it works reasonable with these example models:

- Child type is "Phone" (//simple//) and custom association type is "Phone Entry" (//composite//).
  -> The Phone field is labeled "Phone" then.

- Child type is "Date" (//simple//) and custom association type is "Birthday" (//simple//).
  -> The Date field is labeled "Birthday" then.

- Child type is "Association Type" (a meta type, technically //simple//) and custom association type is "Custom Association Type" (//simple//).
  -> The Association Type field is labeled "Custom Association Type" then.

The latter example shows part of DM's meta model definition. Click an association definition instance to see the effect in the detail panel.

Thank you, JuergeN, for comming up with this label rule!

See #341.
See #779.