Webclient: report error responses to user (#790).
When the RESTClient receives an error response the error is reported to the Webclient user via an alert box. The alert box displays the message and class name of the (Java) exception occurred at server-side, as well as all the cause-exceptions.

CHANGES

RestClient API:

The developer can pass an error handler to the RESTClient by setting the `on_error` property in the config object when instantiating the RESTClient. In cases of an error the RESTClient calls this handler and passes an exception object:
{{{
{
    exception: "<exception class name>",
    message: "...",
    cause: {
        exception: "<exception class name>",
        message: "...",
        cause: {
            exception: "<exception class name>",
            message: "..."
        }
    }
}
}}}
Note: the final cause exception has no `cause` property.

See #790.