RESTClient: handle arbitrary error entity (#790).
The RESTClient handles arbitrary error response entities.

The Webclient's error alert box displays a response entity which is a JSON represented Java exception in a proper layout. Other JSON entities and non-JSON entities are displayed unformatted (as a fallback).

For server-side developers:

To generate a standard 500 (Server Error) response with the JSON representation of a Java exception just throw a RuntimeException.

To generate an error response with a custom status code and/or custom entity throw a WebApplicationException.

CHANGES

The `on_error` RESTClient config property is renamed to `on_request_error`.
The parameter passed to the error callback is the error response, an object with 4 properties:
    content_type -- (string)
    content      -- (string)
    status_code  -- (number)
    status_text  -- (string)

See #790.