Changing passwords work again (#795).
Background: since DM 4.6 the REST API and declarative migrations support a simplified format for create/update requests. The child topic's type_uri must no longer be specified as it is redundant (see #636). The type_uri field is initialized by the server automatically.

The DM 4.6 Webclient made use of the simplified format and did not initialze the topic's `type_uri` in update requests. This broke the working of plugins which register a listener for the **client-side** `pre_update_topic` event as most plugins do work based on the topic's type there.

The Access Control plugin uses the `pre_update_topic` listener to encode a changed password at client-side before it is send to the server.

Now with this fix the topic/assoc passed to the `pre_update_topic`/`pre_update_association` listeners is again guaranteed to have its `type_uri` field initialized.

See #795.