Mobile: improve page panel resize handling (#803).
The width of the resize-handle is increased from 7px to 16px to make it easier to grasp.

When tapped the resize-handle appears visually to provide a hint where to grasp it.

Still, resizing remains very frustrating. Most resize attempts fail.
The various mobile browsers seem to be governed by subtle and opaque rules.

See #803.
See #676.