Enforce per-user file upload quota (#812).
New config property:
{{{
dm4.filerepo.user_quota
}}}
Per-user file upload quota in MB.
Default is 150.

The quota is enforced by the UploadedFile.write() method.
Your plugin should use this method to write an uploaded file to disc.

The disc space occupied by a user is calculated by retrieving all File topics where that user is the creator. Then the size of the underlying file is get by asking the file system (not by the File topic's Size field). Files for which no File exists are not respected in the calculation.

Your plugin should make sure to create the corresponding File topic when writing a file to disc. Using the FilesService's storeFile() method is recommended. Note: POSTing a file to /files service is not mandatory to upload a file. Your plugin can receive an upload, process it in some way, and then possibly call the FilesService's storeFile() method.

See #812.