Fix: files with "+" in path are rendered (#819).
A file whose path contains a "+" character is properly rendered in the detail panel.

BREAKING CHANGES

A file repository path contained in a `/files/` or `/filerepo/` request must be encoded as a URI component, e.g. with Javascript's `encodeURIComponent` function. In particular the "+" character must be encoded ("%2B"). (The `encodeURI` function would not work.)

Plugins that use the Files plugin's client-side `open_upload_dialog(upload_url, callback)` utility method are responsible for encoding the passed URL.

When using a reverse proxy it must be configured to NOT perform any URL decoding before passing the request to the application.

See #819.