Secure file repo (#815).
The service which serves the static file repo files is now secured.
If a `GET /filerepo/{repo_path}` request is not authorized the response is 401 Unauthorized.
Authorization is performed by checking access to the File topic that corresponds to the repo path.
If that File topic is missing the response is 500 Server Error.

Example of a file repo request (multi-user setup):
{{{
GET /filerepo/workspace-1234/images/img1.png
}}}
Here 1234 is the ID of the workspace the file belongs to.

All this applies to the situation when the DM file repo is put under access control, that is when the `dm4.filerepo.per_workspace` config property is set to `true`.

In contrast in a single-user situation, that is when `dm4.filerepo.per_workspace` is `false` (the default), no file repo access control is performed. In this case the existence of a corresponding File topic for each repo file is not a requirement. Consider the DM default configuration, which is "single-user" and regards the entire hard disc as the file repo.

Example of a file repo request (single-user setup):
{{{
GET /filerepo/home/terry/images/img1.png
}}}
No workspace-specific directory appears here.

CHANGES

2 new methods in `de.deepamehta.core.service.accesscontrol.AccessControl`:
{{{
String getUsername(HttpServletRequest request);

String username(HttpSession session);
}}}

See #815.