Adapt Webclient to jQuery UI 1.11.4 (#507, #758).
With jQuery UI 1.11.4 we didn't manage to get scrollable menus with submenus layouted properly.
The submenu appears inside the main menu's container and a vertical scrollbar appears. The submenu is only visible after scrolling to the right.
Fiddling with overflow-x and overflow-y doesn't help. The intended x=visible/y=auto combination is not supported by browsers. See W3C spec below.

With older jQuery UI 1.10.4 we got it working!
The solution is to set the menu's CSS "position" to "static" (and possibly position the menu by the jQuery UI "position" option then). All other position values ("absolute", "relative", "fixed") are not working. CSS "overflow" can just be "auto" then.
With jQuery UI 1.11.4 the "position=static" solution do not work.
So, scrollable menus with submenus are possible with jQuery UI 1.10.4 but not with newer jQuery UI 1.11.4.
All this is true pain!!

W3C spec says:

The computed values of "overflow-x" and "overflow-y" are the same as their specified values, except that some combinations with "visible" are not possible: if one is specified as "visible" and the other is "scroll" or "auto", then "visible" is set to "auto". The computed value of "overflow" is equal to the computed value of "overflow-x" if "overflow-y" is the same; otherwise it is the pair of computed values of "overflow-x" and "overflow-y".

http://www.w3.org/TR/css3-box/#overflow-x

See #507.
See #758.