Config plugin: add "Configure" command (#830).
The Webclient shows a "Configure" command in the topic context menu if configuration definitions have been registered. THe "Configure" command features a submenu which lists all the registered config types. Once a config type is selected the corresponding config topic is (created if not yet exists and then) revealed, ready for being edited in the detail panel.

See #830.