Failed uploads are reported to the user (#815).
If an error occurs while a file upload the error is reported to the user by the means of a dialog box.

CHANGES

Client-side `dm4c` API:

1 new method:
{{{
dm4c.open_error_dialog(server_response)
}}}

Opens an error dialog and renders a server response.

The server response will be rendered as
1) either arbitrary error text or as a chain of exceptions, and
2) the server status.

A chain of exceptions is rendered if the server response meets these conditions: "content_type" is
"application/json" and "content" is the JSON representation of a Java exception (as occurred at server-side):
an object with "exception", "message", and "cause" properties. The "cause" value is again an exception object.
The final exception has no "cause" property.
{{
@param   server_response  The server response to render. An object with 4 properties:
             content_type -- media type of the error content (string)
             content      -- the error content (string)
             status_code  -- the server status code (number)
             status_text  -- the server status text (string)
}}}

See #815.