Drop @ProvidesService annotation (#836).
Sorry, change of plan: the @ProvidesService annotation is dropped after all.

BREAKING CHANGE

New rule for recognizing which service interface a plugin implements: among the interfaces implemented by a plugin the service interface is the one whose name ends with "Service". No annotation is needed. This is more "Convention over Configuration" like.

The service interface can be located anywhere (even in another module), but the convention is for it to reside in the plugin's main package. The package "service" has no special meaning anymore, and, according to the convention, is not used anymore.

See #836.