Introduce server-side preInstall() hook (#843).
The preInstall() hook is triggered //before// the plugin is installed in the database.
The preInstall() hook is triggered each time the plugin starts, just like the init() hook.

The preInstall() hook is useful in particular in conjunction with the new Config service (#830).

CHANGES

Slight change in init() hook: the init() hook is triggered //after// the plugin's event listeners and OSGi service are registered, that is at the very end of the plugin activation phase. Thus at init() time the plugin developer can be shure all the plugin aspects are fully functional.

Formerly the init() hook was triggered after installing the plugin in the database but //before// registering the event listeners and OSGi service.

Overview of the plugin activation phase:
1. invoke preInstall() hook
2. install plugin in the database
   (includes creating plugin topic, running migrations, fire type introduction events)
3. register event listeners
4. register OSGi service
5. invoke init() hook

See #843.