Change standard plugins hot deploy (#847).
For core developers: There is only //one// hot deploy directory.
That is called "bundle-deploy" and is created in the repo root directory.
Every time a standard plugin is build it is automatically copied to the bundle-deploy directory.

Pending: don't copy the deepamehta-test artifact (which is not a bunble) to bundle-deploy.

See #847.