POMs: deepamehta-test is not hot deployed (#847).
INTERNAL CHANGES
(not relevant for plugin developers)

Artifact `deepamehta-test` is derived from `deepamehta-global` (instead of `deepamehta-parent`).
Thus it is not copied to bundle-deplopy (= not hot deployed) as it is no bundle.

Elements moved from `deepamehta-parent` to `deepamehta-global`:
* <dependencyManagement>
* junit, easymock <dependencies>

`deepamehta-parent` version numbers are in sync with the DM distro version number (4.8-SNAPSHOT).
Independent versioning (... 5 6 7) is not reasonable as it changes with every distro release.

2 renamed artifacts:
{{{
deepamehta-global -> deepamehta-project
deepamehta-parent -> standard-module
}}}

See #847.