Introduce "Administration" workspace (#855).
The "Administration" workspace is //collaborative//.
Its owner is the "admin" user.

The config topics ("Login enabled", "Disk quota") are assigned to it.
Thus they are not visible to non-admin users.

A migration that reassigns existing config topics is provided.

The URI of the "Administration" workspace is "dm4.workspaces.administration"

1 new method in `de.deepamehta.core.service.accesscontrol.AccessControl`:
{{{
long getAdministrationWorkspaceId()
}}}

See #855.