Files: use workspace name as Folder Name (#850).
A Folder topic that represents a workspace root directory uses the workspace name as Folder Name.
This is more intuitive as the workspace //ID// have no meaning to the user.

This applies only if per-workspace file repos are activated, that is:
{{{
dm4.filerepo.per_workspace=true
}}}

A migration that transforms existing Folder topics is provided.

Thank you, JuergeN, for coming up with this usability improvement!
:-)

See #850.