Adds "Enabled Sharing Modes" config type (#856).
Adds an "Enabled Sharing Modes" configuration on a per-user(name) basis. It allows individual enabling/disabling of the 5 Workspace Sharing Modes. A user can only use the enabled Sharing Modes when creating a workspace.

A migration to add the default configuration to existing Usernames is provided.

Pending: client-side part.

See #856.