Build: introduce dm4.deploy.dir property (#847).
The developer can keep hers plugin sources at a non-standard location (other than "bundle-dev" resp. "modules-external"). In order to hot (re)deploy the plugin every time it is build the developer must add the property `dm4.deploy.dir` to the plugin's `pom.xml`. The value is a path to the directory where DM copies the plugin jar to once build.

The path can be absolute:
{{{
<properties>
    <dm4.deploy.dir>/home/terry/deepamehta/bundle-deploy</dm4.deploy.dir>
</properties>
}}}

or relative (to the plugin pom.xml):

{{{
<properties>
    <dm4.deploy.dir>${project.basedir}/../../deepamehta/bundle-deploy</dm4.deploy.dir>
</properties>
}}}

The default value for `dm4.deploy.dir` is `../../bundle-deploy`, which is fine in case the plugin sources are kept at the standard location ("bundle-dev" resp. "modules-external").

See #847.
See #846.