Files API: add 2 pathPrefix() methods (#884).
2 new methods in Files API:

{{{
String pathPrefix()
}}}
Returns a prefix that can be used for constructing a repository path.
In case of per-workspace file repos are activated (`dm4.filerepo.per_workspace=true`) the prefix
represents the //current// workspace (e.g. `/workspace-1234`), based on the workspace cookie.
In case of per-workspace file repos are //not// activated an empty string is returned.

{{{
String pathPrefix(long workspaceId)
}}}
Returns a prefix that can be used for constructing a repository path.
In case of per-workspace file repos are activated (`dm4.filerepo.per_workspace=true`) the prefix
represents the //given// workspace (e.g. `/workspace-1234`).
In case of per-workspace file repos are //not// activated an empty string is returned.

See #884.