Files: disbale disk quota by default (#812).
By default the disk quota is disabled (-1). Formerly it was 150 MB.

This is because the DM Standard Distribution's default configuration reflects a single-user installation and a disk quota makes no sense here.

For a multi-user installation on the other hand the administrator is required to adapt several configuration values anyway. There is no reasonable default value for disk quota.

See #812.