Webclient: pluggable composite renderers (#908).
A composite renderer can be registered like a simple renderer:
{{{
dm4c.add_simple_renderer(renderer_uri, renderer)
}}}

Also the renderers `info` and `form` methods are analogous to a simple renderer, instead that `page_model` (as passed by the webclient) is a composite model:
{{{
render_info(page_model, parent_element, level)
render_form(page_model, parent_element, level)
}}}

At the moment the function to process a composite form is still hardcoded.
So, at the moment `render_form()` is not required to return the form processing function.

In a View Configuration you setup a composite renderer by setting //Simple Renderer URI// to the renderer's URI.
So, there is no new view config type added. //Simple Renderer URI// is used for both, simple renderers, and composite renderers.

The wording //Simple Renderer// is about to be changed to //Single Renderer//, in order to contrast it to //Multi Renderer//. A //Single Renderer// renders a single topic, regardless if it is a simple one, or a composite one.

BREAKING CHANGE

The `render_form()` method is now responsible for rendering the field labels too. Thus your renderer have full control about how/wether the labels of the form fields are rendered.

See #908.