Date/Time detail panel layout (#901).
In the detail panel the Month/Day/Year fields as well as the Hour/Minute fields are respectively layouted in a row (instead of on top of each other). Applies to both, info mode and form mode.

This gives a much more pleasant user experience, in particular e.g. for Event topics (part of the Events module) which employ a lot of date/time information.

The layout is defined solely via CSS (see #929).

Example for the Month/Day/Year fields:
{{{#!css
/* === Date === */

.dm4\.datetime\.month,
.dm4\.datetime\.day,
.dm4\.datetime\.year {
    display: inline-block;
}

.dm4\.datetime\.day,
.dm4\.datetime\.year {
    margin-top: 0 !important;
    margin-left: 1em;
}

/* Form */

.dm4\.datetime\.month input,
.dm4\.datetime\.day   input {
    width: 2em !important;
}

.dm4\.datetime\.year input {
    width: 3em !important;
}
}}}

See #901.