Share Address topics for Events (#902).
To assign an existing Address topic (e.g. of an Institution) to an Event:
1. reveal the Address topic
2. associate the Address topic with the Event topic. DM sets the association's type and role types automatically then.

If an Address has been assigned to the Event already that association is deleted automatically. (The Address topic is not.)

Thus you can now share Address topics between e.g. Institutions and Events. You must not re-enter an Event address again if the Event takes place at an Institution for which an Address has been entered already.

However, its not yet perfect:
- The association that is deleted automatically disappears from the canvas only after reloading.
- The newly assigned Address appears in the Event's details only after reloading.

See #902.