Webservice: create() sends directives (#932).
Directives are sent for //create// operations as well.

Note: while the Core Service's `createTopic()` and `createAssociation()` methods still return a `Topic` or an `Association` respectively, the corresponding Webservice method returns a `DirectivesResponse` object. A `DirectivesResponse` wraps a `DeepaMehtaObject` (the created object) and a set of directives.

Consider e.g. the case when an Address is associated with an Event and the event has already another address. DM deletes the former assignment then, and removes it from the canvas as well, if displayed. In this case a `DELETE_ASSOCIATION` directive is sent as a response to a `createAssociation()` call. The DB and Webclient view are in-sync.

The REST API is still compatible. The response of a `POST /core/topic` or `POST /core/association` request is still the created object, but extended with an additional `directives` property (an array):
{{{
{
  id: ...,
  uri: ...,
  type_uri: ...,
  value: ...,
  childs: ...,
  directives: [...]
}
}}}

See #932