Inject ModelFactory into plugins (#933).
The ModelFactory is available in a plugin main file as `mf`.

BREAKING CHANGE

Plugin developers can't create model objects with the `new` operator anymore.
Instead they must use the `ModelFactory`, that is the `mf` object. Example:

Old code:
{{{
new TopicModel(...)
}}}

Since DM 4.8:
{{{
mf.newTopicModel(...)
}}}

The `mf` object is available in the plugin main file and in imperative migrations (just as the `dms` object is).

The `mf` object adheres to the `de.deepamehta.core.service.ModelFactory` interface.

Note: all items in the `de.deepamehta.core.model` package that end with `Model` are interfaces now.
`SimpleValue` on the other hand remains a class and can still be instantiated with `new` (and `IndexMode` remains an enum).

See #933