Webclient: auto-scale detail panel images (#876).
Images displayed in the detail panel never exceeds the panel width.
If the image is larger it is auto-scaled down to fit the detail panel.
This applies to all images, regardless if it originates from CKEditor or an image File topic.

The Webclient applies this rule:
{{{
#page-content img {
    max-width: 100%;
    height: auto;
}
}}}

`height: auto` avoids distortion in case the <img> element has the `height` attribute already set. This is the case when the image was inserted via CKEditor.

Note: the same rule is added to the CKEditor content stylesheet. That is the same auto-scaling is applied while in edit mode.

See #876.