Core API: DeepaMehtaService wording (#941).
BREAKING CHANGES

Core API

1 interface renamed:
{{{
de.deepamehta.core.service.DeepaMehtaService
->
de.deepamehta.core.service.CoreService
}}}

1 object renamed:
{{{
dms -> dm4
}}}

Background: the wording `DeepaMehtaService` was not concise. What would be the DeepaMehtaService? Possibly the wholeness of //all// the DeepaMehta module services. But such an entity does not exist.

`CoreService` on the other hand denotes exactly what it is: the service of the DeepaMehta Core module.

Consequently `dms` is no longer a suitable name. A candidate would be `cs` then. However with the new `dm4` wording an exception was made as this omnipresent object is of central importance for every DeepaMehta developer.

See #941.