Core API: add findChildTopic() (#941).
1 new method in `Topic` interface:
{{{
Topic findChildTopic(String topicTypeUri)
}}}

Searches this topic's child topics for a topic of the given type.
Only the child topics which are already loaded into memory are searched; the DB is not accessed.

The first topic found is returned, according to a depth-first search.
For multiple-value childs the first topic is returned.
If the given type matches this topic directly it is returned immediately.

The search is driven by this topic's type definition. That is child topics which do not adhere
to the type definition are not found.

Returns the found topic, or `null`.

See #941.