Core: add privileged change-password call (#934).
1 new method in Core `AccessControl`:
{{{
void changePassword(Credentials cred)
}}}

Changes the password of an existing user account. (The username can't be changed.)
An user account with the given username must exist.
This is a privileged method: it works also if the respective user is not logged in.

The DM standard distro provides no REST API for this method.
It's up to a 3rd-party plugin to establish a proprietary authorization procedure before calling this method.

Hint: there are 2 approaches to instantiate a `Credentials` object:

* Instantiation through JAX-RS: transport a JSON object in the message body ...
  {{{
  {
    username: "john",
    password: "-SHA256-encoded-password"    # the -SHA256- prefix is mandatory
  }
  }}}

  ... and receive it in your resource method as an entity parameter:
  {{{
  import de.deepamehta.core.service.accesscontrol.Credentials;

  @GET/PUT
  @Path("/your-path")
  @Transactional
  public void yourResourceMethod(Credentials cred) {
  	...
  }
  }}}

* Instantiating yourself:
  {{{
  new Credentials("john", "plaintext-password")
  }}}

  Note that the password is expected as plain text in this case.
  SHA256-encoding is done by DM.

See #934.