Core: username/email privileged calls (#934).
2 new methods in Core's `AccessControl`:

{{{
String getUsername(String emailAddress);
}}}

Returns the username for the given email address.

The username is determined by traversing from the Email Address topic along a
`org.deepamehta.signup.user_mailbox` association.

This is a privileged method, it bypasses the access control system.

@throws  RuntimeException
if no such Email Address topic exists in the DB, or
if more than one such Email Address topics exist in the DB, or
if the Email Address topic is not associated to a Username topic.

{{{
String getEmailAddress(String username);
}}}

Returns the email address for the given username.

The email address is determined by traversing from the Username topic along a
`org.deepamehta.signup.user_mailbox` association.

This is a privileged method, it bypasses the access control system.

@throws  RuntimeException
if no such Username topic exists in the DB, or
if the Username topic is not associated to an Email Address topic.

See #934.