Core Access Control: revise username/email (#934).
2 Core `AccessControl` calls are revised:

{{{
String getUsername(String emailAddress)
}}}

In case of more than one Email Address topics exist, the one that is assigned to a Username topic is returned.
Formerly an exception was thrown.

{{{
boolean emailAddressExists(String emailAddress)
}}}

True is returned only if the email address is actually assigned to a Username topic.

See #934.