Files module: add "Download" button (#661).
A File topic features a "Download" button in the detail panel.
Pressing it causes the browser to download (and open/save) the file in its ideosyncratic way.

REST API

A `/filerepo/` request can be equipped with a `download` query parameter.
If present a `Content-Disposition: attachment` header will be added to the response.
This causes the browser to download the requested file (instead of rendering it).

Example:
{{{
GET /filerepo//my-dir/my-song.mp3?download
}}}

See #661.