Detail panel: custom assoc type as label (#1023).
The name of a Custom Assoc Type is used for labeling a topic in the detail panel also in case the topic is COMPOSITE, provided the Custom Assoc Type is SIMPLE.

Thus, e.g. "From", "To", "Date of Birth" Date/Time fields (from Events and Contacts modules respectively) are labeled properly, without resorting to CSS pseudo elements.

See #1023.