Core: fix per-user type projection, pt. 1 (#1021)
When creating a type the type topic is instantiated explicitly on a `TopicModel` (which is freshly created from the `TypeModel`). Creating the type topic from the `TypeModel` directly would fail as topic creation implies topic instantiation, and due to the polymorphic `instantiate()` method a `Type` object would be instantiated (instead a `Topic` object). But instantiating a type newly implies per-user type projection, that is removing the assoc defs not readable by the current user. But at the time the type topic is stored in the DB its assoc defs are not yet stored, and the readability check would fail.

See #1021.