Core: fix per-user type projection, pt. 2 (#1021)
Type updates are performed on the //internal// (= **kernel**) type model instead on the userland's type model. Performing an update on the **userland**'s type model would have no effect, as it is transient. The userland's type model is always a //cloned// and filtered (= "projected") version of a kernel type model which is created on-the-fly each time a specific user requests it.

See #1021.