Webclient start: topicmap ID validity (#1017).
On Webclient start the ID of the topicmap initially shown is cheked for validity.
The initial topicmap ID is obtained from two sources (in this order):
* The browser URL path (/topicmap/<id>/topic/<id>)
* From the `dm4_topicmap_id` browser cookie
If both IDs are not present resp. not valid, the ID is obtained from the 1st item of the Topicmap menu as a fallback.

An ID is regarded valid if all applies:
* The object exists in the DB and is actually a topic (not an assoc)
* The topic is READable by the current user
* The topic is actually a Topicmap (not a Note or something else)

So, the Webclient starts properly even if another DB is present meanwhile, e.g. after a DB reset.
The stale browser cookie is fixed.

See #1017.