Webclient: fix topicmap initialization (#1017).
There was another intricate edge case: the Topicmaps plugin might regard an ID as valid topicmap ID although it is not. This happens if a stale Topicmap topic is obtained from the browser cache. This happens if the GET request's condition (If-Modified-Since) is not satisfied and the response is 304 Not Modified. For this to happen it is sufficient if the pretended topicmap ID (e.g. from a stale cookie) refers to *some* existing DB object (topic or association!) whose modification timestamp is "old enough"! This case is not detectable by the Topicmaps plugin at init(1) stage and will only be detected by the Workspaces plugin (at init(2) stage). The Topicmaps plugin will finally obtain a valid topicmap ID through a fallback at init(3) stage.

See #1017.