Core API: getType() has implicit READ perm (#941).
1 `DeepaMehtaObject` method has changed semantics:
{{{
DeepaMehtaType getType()
}}}

No access control is performed as //Implicit READ permission// applies: if a user has
READ access to an object she has READ access to its type as well.

Note: if the user would have no READ access to the object the DeepaMehta Core would not
instantiate it in the first place, but throw an <code>AccessControlException</code>.

See #941.