Fix "suppress redundant topics" for assoc (#1070).
In the detail panel's "What's related" list: redundant topics are suppressed from being displayed also when an association is selected.

BREAKING CHANGE

1 changed URI in Webclient REST API:
{{{
/webclient/topic/{id}/related_topics
->
/webclient/object/{id}/related_topics
}}}

Works for both now, topics and associations.

This request returns an object's related topics, but suppressing the ones which are child topics according to the object's type definition.

See #1070.